Fix stm32f3 README about what module to use for STM32F318 and STM32F378
STM32F318 is documented in RM0366 and not in RM0316 so it should use
the stm32f301 module.

STM32F378 is documented in RM0313 and not in RM0316 so it should use
the stm32f373 module.

Also the st.com links for stm32f373 and stm32f3x8 where inverted.