rename DBG to DBGMCU
Some of the families name this device DBGMCU and others DBG, even
though the Reference Manual for all thoses families are describing
this device in the exact same way.
It is presented in a DBG chapter which contains various kind of
devices such as ETM or TPIU as well as DBGMCU, each with their own
register set.
So for consistency across families, and for better matching of the
reference manual, I propose that we name this device DBGMCU in all
families.