stm32h7x3: add SPI peripheral documentation
Two things are not covered:

The SVD has all six SPI peripherals the same and derived from SPI1.
In reality only SPI123 are as documented. SPI456 have 16 bit wide TX and
RX registers and DSIZE is limited accordingly.

Furthermore, there should be explicit byte and half-word access to the
TX and RX data registers. This is how the access-transfer packing ratio
is determined.