stm32h7: Rename stm32h7x3 into a total of 4 devices
stm32h743/753 with revision v versions

The resulting patched SVDs for stm32h753 (not rev v) is identical to
the previous output for stm32h7x3. The stm32h743 is almost identical,
less the CRYP peripheral and interrupts.

A less awful alternative to the symbolic links would be appreciated...

ST have released v1.5 SVDs (we currently have v1.4). However the only
change in the release is to remove/hide the 7a3x / 7b3x SVDs. So the
archive is not updated in this PR.