stm32h7: dualcore based on stm32h7x5 instead
As far as I can tell the H7x7 and H7x5 are the same die, but binned
for different voltage-temperature ranges (and different
packages). They share RM0399 with no mention of differences.

The `stm32h7x5` SVD has a different set of errors to the `stm32h7x7` SVD,
but overall it's not quite so bad. Therefore use this instead.

I used the diff between the two SVDs to highlight errors and resolved
according to the RM. The previous problems with ADC and DMAMUX are
resolved.