Fix stm32f3 README about what module to use for STM32F318 and STM32F378
STM32F318 is documented in RM0366 and not in RM0316 so it should use
the stm32f301 module.

STM32F378 is documented in RM0313 and not in RM0316 so it should use
the stm32f373 module.

This was already applied in ef858ad but was missing the
stm32_part_table.yaml change so it was overwritten when
makecrates.py was executed again.