sai: sai_v1 patch and sai peripheral apply to H7. Fix FSDEF
SAI peripheral applies to the H7 also. Note that we introduce the very
slightest difference with RM0433 (single core 743/753) by setting
`[AB]CR1.NOMCK` -> `[AB]CR1.NODIV`. Both RM0399 (later dual core
parts) and `sai/sai.py` use NODIV. The description in `sai/sai.py` has
text from both the old and new definitions, and it appears the
functionality of the bit is unchanged.

[AB]FRCR.FSDEF is marked as read only in the SVD and the register
diagram in the RM. However, in the prose text in the RM is is clearly
noted as read-write "This bit is set and cleared by software." (RM0399
rev2 page 2458).