Add RAMECC registers across the H7 family
RAMECC was 'added' to the single core parts in RM0433 Rev 7. For the
dual core parts there are definitions, but they are so terrible it's
worth starting over. (there's registers with the wrong address and wrong fields)

The RAMECC2 and RAMECC3 registers are derived from RAMECC1. RAMECC3
only has two useful monitor channels, but there seems to be no harm in
accessing the remaining channels.

For a use case, there is at least one set of commercial spaceflight
hardware using the H7: (no affiliation, no endorsement intended)

- [Nanoavionics OBC](https://nanoavionics.com/cubesat-components/cubesat-on-board-computer-main-bus-unit-satbus-3c2/)