stm32f4: rename I2C4/I2CFMP1 to FMPI2C1
The Fast Mode Plus I2C peripheral in STM32F4 devices have some quite
inconsistent names. Here is a summary of the various names I found
in the SVD and the Reference Manuals.

| Dev  | SVD Peri | SVD RCC         | SVD SYSCFG | RM Peri  | RM RCC  | RM SYSCFG |
|------|----------|-----------------|------------|----------|---------|-----------|
| F410 | I2C4     | I2C4            | FMPI2C1    | I2C4 FM+ | I2C4    | FMPI2C4   |
| F412 | I2C4     | I2C4            | I2C4       | I2CFMP1  | I2CFMP1 | I2CFMP1   |
| F413 | FMPI2C   | I2C4/I2CFMP1    | I2C4       | I2CFMP1  | I2CFMP1 | I2CFMP1   |
| F446 | Missing  | I2CFMP1/FMPI2C1 | Missing    | Missing  | FMPI2C1 | FMPI2C1   |

The peripheral documentation section is named FMPI2C in all Reference
Manuals, so it seems obvious that it is the prefix we should use.
With all the above, I think FMPI2C1 is the most sensible name.