h7: Rename USBxOTGHSEN -> USBxOTGEN, USBxOTGULPIHSEN -> USBxULPIEN
The effect of this PR is different between the H7 groupings:

highmemory (RM0455): This change brings us in line with the RM, fixing a mistake
on our part

singlecore/dualcore (RM0433/RM0399): Breaks with the RM, instead preferring a
more consistent scheme that matches RM0455. All other peripherals enable/rests
are in the form name{en,rst} , so having the inconsistent name{hsen,rst} is a pain.

Note: I'm not sure the existance of USB2ULPIEN makes any sense (the USB2
peripheral has no ULPI), but it is RM0433/RM0399.