Restores the correct definition for IWDG and WWDG for the g4 family.
The svd files for the g4 family incorrectly swap the names of IWDG and WWDG,
such that the registers and definitions of IWDG appear under the WWDG peripheral
and vice-versa.

A patch that deletes IWDG and WWDG and then restores them to the correct values
was added.

The patch was applied to the whole g4 family of devices.