stm32wb: ADC -> ADC1 and add ADC_COMMON
As is commonly done, use "ADC1" even when there's only one ADC.  This
makes it not only more self consistent, (interrupt names were for ADC1)
but also more consistent with sibling targets like STM32L4.

Also, like the L4, the WB svd is missing the ADC_Common registers.  Add
them back in, but note that there's less bits in the common registers.
Note that CCR _is_ defined, but as part of the core block, so we have to
delete it from there as well.

Signed-off-by: Karl Palsson <karlp@etactica.com>