Fix ADC disable
`ADDIS`, as `ADEN` is a signal to the ADC peripheral to change it's
state. Both signals have to be "1" to change the state.

In case of `ADDIS` it will be only reset to "0", after the state has
successfully changed.

The procedure of disabling is the same between `stm32f303xc` (RM0316
15.3.10), where I've noticed the bug, and `stm32h7x3` (RM0443 25.4.9),
for which this code was initially committed, so I assume this affects
every device using the `adc_v3.yaml`.