Fix interrupt definitions for UARTs 7,8,9,10
UARTs 7,8,9,10 have incorrect interrupt definitions.  For 7,8 copy the
definitions from f469 instead.  Switch to using the f469 definitions for
UART4,5 too (these are identical to the f405 ones, but in the f405 svd,
UART7,8 also have incorrect interrupt definitions, so switch all UARTs
to copy from  f469 for for simplicity).